backend/dogback/requirements.txt
absl-py==1.3.0
asgiref==3.5.2
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==37.0.2
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==3.2
django-admin-honeypot==1.1.0
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-haystack==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
environ==1.0
executing==1.2.0
flatbuffers==22.11.23
fonttools==4.33.3
gast==0.5.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
idna==3.3
importlib-metadata==5.1.0
ipython==8.6.0
jedi==0.18.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
numpy==1.22.4
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
traitlets==5.5.0
trio==0.20.0
trio-websocket==0.9.2
types-cryptography==3.3.23
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.2.0
Whoosh==2.7.4
wrapt==1.14.1
wsproto==1.1.0
zipp==3.11.0
